cat_spotter/requirements.txt
streamlit==1.38.0
tensorflow==2.15.0
pillow==10.3.0
requests==2.32.3
numpy==1.26.4